Move vote federation change nested test class to its own file
Huge refactor to get rid of duplicated code

Minor refactor

Minor refactors

Renaming, refactors. Add missing test case.

Minor fix

Reorder, refactor

Fix typo, refactor

Rename

Create private method to get rid of duplicated code
